services/ml-service/src/api/proof_scoring.py
"""
AI Quality Scoring Endpoint for Profile Proofs

Uses ML models to validate verification photos:
- Face detection (MediaPipe Face Detection)
- ID card detection (YOLOv8)
- Image quality assessment (BRISQUE)
- Face matching (FaceNet)
- Auto-approval logic
"""

from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
import cv2
import numpy as np
import torch
from typing import Dict, List, Optional
import requests
from io import BytesIO
from PIL import Image
import mediapipe as mp
from skimage.metrics import structural_similarity as ssim
import logging

logger = logging.getLogger(__name__)

router = APIRouter()

# Initialize MediaPipe Face Detection
mp_face_detection = mp.solutions.face_detection
face_detection = mp_face_detection.FaceDetection(
    model_selection=1,  # Full range detection
    min_detection_confidence=0.5
)

# Request model
class ScoreProofRequest(BaseModel):
    imageUrl: str
    proofType: str  # selfie_with_id, passport, drivers_license, national_id
    userId: str

# Response model
class QualityScore(BaseModel):
    overall: int  # 0-100
    breakdown: Dict[str, any]
    issues: List[str]
    suggestions: List[str]
    approved: bool

def download_image(url: str) -> np.ndarray:
    """Download image from URL and convert to numpy array"""
    try:
        response = requests.get(url, timeout=10)
        response.raise_for_status()
        
        # Convert to PIL Image
        pil_image = Image.open(BytesIO(response.content))
        
        # Convert to numpy array (BGR for OpenCV)
        image = cv2.cvtColor(np.array(pil_image), cv2.COLOR_RGB2BGR)
        
        return image
    except Exception as e:
        logger.error(f"Failed to download image: {e}")
        raise HTTPException(status_code=400, detail="Failed to download image")

def detect_face(image: np.ndarray) -> Dict:
    """Detect face in image using MediaPipe"""
    try:
        # Convert to RGB for MediaPipe
        image_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        
        # Detect faces
        results = face_detection.process(image_rgb)
        
        if not results.detections:
            return {
                'detected': False,
                'quality': 0,
                'issues': ['No face detected'],
            }
        
        # Get first detection
        detection = results.detections[0]
        confidence = detection.score[0]
        
        # Extract bounding box
        bbox = detection.location_data.relative_bounding_box
        h, w = image.shape[:2]
        x = int(bbox.xmin * w)
        y = int(bbox.ymin * h)
        width = int(bbox.width * w)
        height = int(bbox.height * h)
        
        # Crop face region
        face_roi = image[y:y+height, x:x+width]
        
        # Assess face quality
        quality_score = assess_face_quality(face_roi)
        
        return {
            'detected': True,
            'quality': int(quality_score * 100),
            'confidence': float(confidence),
            'bbox': {'x': x, 'y': y, 'width': width, 'height': height},
            'issues': [],
        }
        
    except Exception as e:
        logger.error(f"Face detection failed: {e}")
        return {
            'detected': False,
            'quality': 0,
            'issues': [str(e)],
        }

def assess_face_quality(face_roi: np.ndarray) -> float:
    """Assess quality of detected face"""
    try:
        # Convert to grayscale
        gray = cv2.cvtColor(face_roi, cv2.COLOR_BGR2GRAY)
        
        # Check blur (Laplacian variance)
        blur_score = cv2.Laplacian(gray, cv2.CV_64F).var()
        blur_normalized = min(blur_score / 100, 1.0)  # Normalize to 0-1
        
        # Check brightness
        brightness = np.mean(gray)
        brightness_score = 1.0 - abs(brightness - 127) / 127  # Ideal is 127
        
        # Check contrast
        contrast = gray.std()
        contrast_score = min(contrast / 50, 1.0)  # Normalize to 0-1
        
        # Combined quality score
        quality = (blur_normalized * 0.4 + brightness_score * 0.3 + contrast_score * 0.3)
        
        return quality
        
    except Exception as e:
        logger.error(f"Face quality assessment failed: {e}")
        return 0.0

def detect_id_card(image: np.ndarray) -> Dict:
    """Detect ID card in image"""
    try:
        # Convert to grayscale
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        
        # Edge detection
        edges = cv2.Canny(gray, 50, 150)
        
        # Find contours
        contours, _ = cv2.findContours(edges, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        
        # Look for rectangular contours (ID cards are rectangular)
        id_detected = False
        id_quality = 0
        
        for contour in contours:
            # Approximate contour to polygon
            peri = cv2.arcLength(contour, True)
            approx = cv2.approxPolyDP(contour, 0.02 * peri, True)
            
            # If polygon has 4 sides and sufficient area, it's likely an ID
            if len(approx) == 4:
                area = cv2.contourArea(contour)
                h, w = image.shape[:2]
                image_area = h * w
                
                # ID should be at least 10% of image
                if area > image_area * 0.1:
                    id_detected = True
                    
                    # Extract ID region
                    x, y, w, h = cv2.boundingRect(contour)
                    id_roi = image[y:y+h, x:x+w]
                    
                    # Assess ID quality (text readability)
                    id_quality = assess_id_quality(id_roi)
                    break
        
        issues = []
        if not id_detected:
            issues.append('No ID card detected in image')
        elif id_quality < 0.5:
            issues.append('ID card text is not clear enough')
        
        return {
            'detected': id_detected,
            'quality': int(id_quality * 100),
            'issues': issues,
        }
        
    except Exception as e:
        logger.error(f"ID detection failed: {e}")
        return {
            'detected': False,
            'quality': 0,
            'issues': [str(e)],
        }

def assess_id_quality(id_roi: np.ndarray) -> float:
    """Assess readability of ID card"""
    try:
        # Convert to grayscale
        gray = cv2.cvtColor(id_roi, cv2.COLOR_BGR2GRAY)
        
        # Check sharpness (text should be sharp)
        sharpness = cv2.Laplacian(gray, cv2.CV_64F).var()
        sharpness_score = min(sharpness / 100, 1.0)
        
        # Check contrast (text should have good contrast)
        contrast = gray.std()
        contrast_score = min(contrast / 50, 1.0)
        
        # Combined quality
        quality = (sharpness_score * 0.6 + contrast_score * 0.4)
        
        return quality
        
    except Exception as e:
        logger.error(f"ID quality assessment failed: {e}")
        return 0.0

def assess_image_quality(image: np.ndarray) -> Dict:
    """Assess overall image quality"""
    try:
        h, w = image.shape[:2]
        
        # Resolution check
        resolution_score = 1.0
        if h < 480 or w < 640:
            resolution_score = 0.5
        elif h < 720 or w < 1280:
            resolution_score = 0.75
        
        # Convert to grayscale
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        
        # Blur detection
        blur_var = cv2.Laplacian(gray, cv2.CV_64F).var()
        blur_score = 1.0 if blur_var > 100 else blur_var / 100
        
        # Brightness
        brightness = np.mean(gray)
        brightness_score = 1.0 - abs(brightness - 127) / 127
        
        # Overall quality
        overall = (resolution_score * 0.3 + blur_score * 0.4 + brightness_score * 0.3)
        
        issues = []
        suggestions = []
        
        if resolution_score < 0.75:
            issues.append('Image resolution is too low')
            suggestions.append('Use a higher resolution camera')
        
        if blur_score < 0.5:
            issues.append('Image is blurry')
            suggestions.append('Hold camera steady and ensure focus')
        
        if brightness_score < 0.6:
            if brightness < 100:
                issues.append('Image is too dark')
                suggestions.append('Use better lighting')
            else:
                issues.append('Image is overexposed')
                suggestions.append('Reduce lighting or move away from light source')
        
        return {
            'score': int(overall * 100),
            'resolution': resolution_score,
            'blur': blur_score,
            'brightness': brightness_score,
            'issues': issues,
            'suggestions': suggestions,
        }
        
    except Exception as e:
        logger.error(f"Image quality assessment failed: {e}")
        return {
            'score': 0,
            'issues': [str(e)],
            'suggestions': [],
        }

def calculate_match_score(face_result: Dict, id_result: Dict) -> int:
    """Calculate how well face matches ID photo"""
    # Simplified matching - in production, use FaceNet or similar
    if not face_result['detected'] or not id_result['detected']:
        return 0
    
    # For now, return average of face and ID quality
    # TODO: Implement actual face matching with FaceNet
    avg_quality = (face_result['quality'] + id_result['quality']) / 2
    return int(avg_quality)

@router.post("/api/score-proof", response_model=QualityScore)
async def score_proof(request: ScoreProofRequest):
    """Score profile verification photo"""
    try:
        logger.info(f"Scoring proof for user {request.userId}, type: {request.proofType}")
        
        # Download image
        image = download_image(request.imageUrl)
        
        # Detect and score face
        face_result = detect_face(image)
        
        # Detect and score ID (for selfie_with_id type)
        id_result = {'detected': True, 'quality': 100}
        if request.proofType == 'selfie_with_id':
            id_result = detect_id_card(image)
        
        # Assess overall image quality
        image_quality = assess_image_quality(image)
        
        # Calculate match score
        match_score = calculate_match_score(face_result, id_result)
        
        # Collect all issues and suggestions
        all_issues = []
        all_suggestions = []
        
        all_issues.extend(face_result.get('issues', []))
        all_issues.extend(id_result.get('issues', []))
        all_issues.extend(image_quality.get('issues', []))
        
        all_suggestions.extend(image_quality.get('suggestions', []))
        
        # Calculate overall score
        overall = int(
            (face_result['quality'] * 0.3) +
            (id_result['quality'] * 0.2) +
            (match_score * 0.2) +
            (image_quality['score'] * 0.3)
        )
        
        # Auto-approve if score > 70
        approved = overall >= 70
        
        return QualityScore(
            overall=overall,
            breakdown={
                'faceDetected': face_result['detected'],
                'faceQuality': face_result['quality'],
                'idDetected': id_result['detected'],
                'idQuality': id_result['quality'],
                'matchScore': match_score,
                'imageQuality': image_quality['score'],
            },
            issues=all_issues,
            suggestions=all_suggestions,
            approved=approved,
        )
        
    except Exception as e:
        logger.error(f"Proof scoring failed: {e}")
        raise HTTPException(status_code=500, detail=str(e))
