services/ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML Libraries
numpy==1.26.3
scikit-learn==1.4.0
torch==2.1.2
torchvision==0.16.2
onnxruntime==1.16.3
transformers==4.36.2

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
scikit-image==0.22.0
pillow==10.2.0

# Supabase Client
supabase==2.3.4

# Redis (Feature Store + Cache)
redis==5.0.1
hiredis==2.3.2

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
orjson==3.9.12
requests==2.31.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
